Avoid concatenate3 in overlap and rechunk
