+++ dask/array/core.py
@@ -54,10 +54,10 @@ from dask.base import (
 from dask.blockwise import blockwise as core_blockwise
 from dask.blockwise import broadcast_dimensions
 from dask.context import globalmethod
-from dask.core import quote
+from dask.core import quote, reshapelist
 from dask.delayed import Delayed, delayed
 from dask.highlevelgraph import HighLevelGraph, MaterializedLayer
-from dask.layers import ArrayBlockIdDep, ArraySliceDep, ArrayValuesDep, reshapelist
+from dask.layers import ArrayBlockIdDep, ArraySliceDep, ArrayValuesDep
 from dask.sizeof import sizeof
 from dask.typing import Graph, Key, NestedKeys
 from dask.utils import (
@@ -5416,6 +5416,11 @@ def stack(seq, axis=0, allow_unknown_chunksizes=False):
     graph = HighLevelGraph.from_collections(name, layer, dependencies=seq2)
 
     return Array(graph, name, chunks, meta=meta)
+
+
+def concatenate_shaped(arrays, shape):
+    shaped = reshapelist(shape, arrays)
+    return concatenate3(shaped)
 
 
 def concatenate3(arrays):
