+++ dask/array/rechunk.py
@@ -22,7 +22,7 @@ from tlz import accumulate
 from dask import config
 from dask._task_spec import Alias, Task, TaskRef, parse_input
 from dask.array.chunk import getitem
-from dask.array.core import Array, concatenate3, normalize_chunks
+from dask.array.core import Array, concatenate_shaped, normalize_chunks
 from dask.array.utils import validate_axis
 from dask.array.wrap import empty
 from dask.base import tokenize
@@ -736,7 +736,12 @@ def _compute_rechunk(x, chunks):
         if all(d == 1 for d in rec_cat_arg.shape):
             x2[key] = Alias(key, rec_cat_arg.flat[0])
         else:
-            x2[key] = Task(key, concatenate3, parse_input(rec_cat_arg.tolist()))
+            x2[key] = Task(
+                key,
+                concatenate_shaped,
+                parse_input(list(rec_cat_arg.flatten())),
+                subdims1,
+            )
 
     del old_blocks, new_index
