+++ dask/core.py
@@ -4,6 +4,8 @@ from collections import defaultdict
 from collections.abc import Collection, Iterable, Mapping, MutableMapping
 from typing import Any, Literal, TypeVar, cast, overload
 
+import toolz
+
 from dask._task_spec import DependenciesMapping, convert_legacy_graph, execute_graph
 from dask.typing import Graph, Key, NoDefault, no_default
 
@@ -525,3 +527,16 @@ def quote(x):
     if istask(x) or type(x) is list or type(x) is dict:
         return (literal(x),)
     return x
+
+
+def reshapelist(shape, seq):
+    """Reshape iterator to nested shape
+
+    >>> reshapelist((2, 3), range(6))
+    [[0, 1, 2], [3, 4, 5]]
+    """
+    if len(shape) == 1:
+        return list(seq)
+    else:
+        n = int(len(seq) / shape[0])
+        return [reshapelist(shape[1:], part) for part in toolz.partition(n, seq)]
